+++ app/src/dhisUITesting/java/org/dhis2/usescases/main/program/ProgramAnimation.kt
@@ -1,0 +1,10 @@
+package org.dhis2.usescases.main.program
+
+import android.graphics.drawable.GradientDrawable
+
+class ProgramAnimation {
+
+    fun initBackdropCorners(gd: GradientDrawable) {}
+
+    fun reverseBackdropCorners(gd: GradientDrawable) {}
+}
